+++ packages/screens/Settings/components/initProfile.ts
@@ -1,0 +1,44 @@
+import {ProfileDataType} from "./UserProfileModal";
+
+export const initProfileData: ProfileDataType[] = [
+  {
+    id: 1,
+    content: "Social Feed",
+  },
+  {
+    id: 2,
+    content: "Albums",
+  },
+  {
+    id: 3,
+    content: "NFTs",
+  },
+  {
+    id: 4,
+    content: "Activity",
+  },
+  {
+    id: 5,
+    content: "Succeed Quests",
+  },
+  {
+    id: 6,
+    content: "Pathwar Challenges",
+  },
+  {
+    id: 7,
+    content: "Gig Services",
+  },
+  {
+    id: 8,
+    content: "Governance Votes",
+  },
+  {
+    id: 9,
+    content: "Putted NFT to Rioters Footer",
+  },
+  {
+    id: 10,
+    content: "Shared servers",
+  },
+]
